FEAT(put pixel): added function; REFACTOR(types): made obj types simpler

+++ includes/ft_engine/mlx_interface.h
@@ -7,6 +7,6 @@
 typedef struct s_main		t_main;
 typedef union u_uint_color	t_uint_color;
 
-void	put_pixel(size_t x, size_t y, t_uint_color color, uint32_t *pixels);
+void	put_pixel(float x, float y, t_uint_color color, uint32_t *pixels);
 
 #endif //MLX_INTERFACE_H
